refactor: Move config store to global folder
